Updates
Updated the app.py and Post_page (individual).

for the app.py I changed some of the mentions of "Single_pots" to the proper file name "Post_Page"

I've fixed tons of bugs with the Post_Page itself. and made sure the comment section / buttons are prepared for the backend.

I've added a section where the code goes in as a text and is displayed on the html as a snippet of code. This was done using the highlight.js framework that was imported into the base html